Addess comments and add delete mapping
